requirements.txt
requests
waveshare_epaper
Pillow
opencv-python
numpy
gradio
argparse